Outline shader Updated for control via Image UI component
The color is now controlled via the Color variable for <Image> component. In script, using UnityEngine.UI and accessing the <Image> component on the Meter Background gameobject of Super Bar element , changing the alpha value of the color variable set in that component will edit it. Current color is E0FF00, No need to update the material itself (except maybe for the outline size if needed)